Try to support ANSI PCBs
I don't have an ANSI PokerII, so this commit needs to be tested by someone
with one. The ISO PCB has keys near the diodes D68 and D69 which i haven't
found on the matrix of the github user tywtyw2002
(https://github.com/erichkeane/poker2firmwarehacking/issues/3).
Instead the key at diode D14 is populated.
D14 was added to the keymap so that ISO and ANSI boards
can be supported by one keymap.