Added kbtflash.py
kbtflash.py can be used to flash the PokerII and also the KBT Pure Pro.
It was originally developed by atalax and released in a gist of him.
There is only a small modification by me, which changes the seventh
byte in the first packet when flashing.

The Pure Pro has the product-id 0x671 and wants to see the byte 0x8c
while the PokerII has the product-id 0x611 and wants the byte 0xec.

Let's wait whether "product-id & 0xff ^ 0xfd" is the correct
way to generate the seventh byte when someone else captures the
usb packets when flashing the next board.