Basic suppport for Drevo Calibur
This is only meant as a proof of concept for the
Nuvoton NUC123 support. You probably don't want
to run it on your keyboard, because there's currently
no support for the LEDs or bluetooth.

But if you do, please try to save the LDROM by
using hid_listen and the Caps Lock+Right CTRL
function mapping of the default keymap.

It can be compiled with "make PROJECT=ty71m".
A debugger is needed to flash the binary.